added delete method

+++ Calender-app/app/src/main/java/com/t3421/calenderapp/Event.java
@@ -15,7 +15,7 @@ public class Event {
     private int day;
     private int year;
     private int month;
-    private long occurenceId;
+    private int occurenceId;
     private String eventName;
     private String eventDetails;
     private String occurance;
@@ -132,9 +132,9 @@ public class Event {
 
     public void setColor(String color) { this.color = color; }
 
-    public  long getOccurenceId(){return occurenceId;}
+    public  int getOccurenceId(){return occurenceId;}
 
-    public void setOccurenceId(long occurenceId){this.occurenceId = occurenceId;}
+    public void setOccurenceId(int occurenceId){this.occurenceId = occurenceId;}
 
     @Override
     public String toString(){
